requirements.txt
asgiref==3.8.1
channels==4.1.0
distlib==0.3.8
Django==5.1
djangorestframework==3.15.2
filelock==3.15.4
mysqlclient==2.2.4
phonenumbers==8.13.44
platformdirs==4.2.2
PyMySQL==1.1.1
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.26.3
